QA-Frontend/requirements.txt
requests==2.18.1
PYYaml==3.12
Flask==0.11.1
Flask-BasicAuth==0.2.0
CherryPy==10.1.1